+++ source/newtonian/two_dimensional/simple_cell_updater.cpp
@@ -35,8 +35,8 @@ namespace {
       0.5*ScalarProd(res.velocity,res.velocity);
 	res.tracers.reserve(extensive.tracers.size());
 	for (size_t i = 0; i < extensive.tracers.size(); ++i)
-		res.tracers.insert(pair<string,double>((extensive.tracers.begin() + i)->first ,
-			(extensive.tracers.begin() + i)->second/extensive.mass));
+	  res.tracers.insert(pair<string,double>((extensive.tracers.begin() + static_cast<int>(i))->first ,
+						 (extensive.tracers.begin() + static_cast<int>(i))->second/extensive.mass));
     res.pressure = eos.de2p
       (res.density,
        energy,
